Began learning JScript.NET
Began learning JScript.NET. The first example I worked on was creating a
windows form with a picturebox that spans the entire size and drawing
rectangles throughout the width of it. Needless to say, there's much
more to learn. Trying to maintain the C# next-line bracket syntax since
it's in the .NET framework.